Split flood-fill algorithm out into a separate source file
